move sync deduplicator + add readme + add into dir
