+++ creator-node/src/serviceRegistry.js
@@ -2,7 +2,7 @@ const redisClient = require('./redis')
 const { ipfs, ipfsLatest } = require('./ipfsClient')
 const BlacklistManager = require('./blacklistManager')
 const MonitoringQueue = require('./monitors/MonitoringQueue')
-const { SnapbackSM } = require('./snapbackSM')
+const { SnapbackSM } = require('./SnapbackStateMachine/snapbackSM')
 const AudiusLibs = require('@audius/libs')
 const config = require('./config')
 const URSMRegistrationManager = require('./services/URSMRegistrationManager')
